Introduce contact owner
