+++ packages/server/customer-os-api/graph/resolver/test_queries/create_contact_with_owner.txt
@@ -1,0 +1,20 @@
+mutation CreateContact($ownerId: ID) {
+contact_Create(
+    input: {
+      title: MR
+      firstName: "first"
+      lastName: "last"
+      ownerId: $ownerId
+    }
+  ) {
+    id
+    title
+    firstName
+    lastName
+    owner {
+        id
+        firstName
+        lastName
+    }
+  }
+}
