+++ packages/server/customer-os-api/graph/schemas/contact.graphqls
@@ -104,6 +104,9 @@ type Contact implements ExtensibleEntity & Node {
     Definition of the contact in customerOS.
     """
     definition: EntityDefinition @goField(forceResolver: true)
+
+    "Contact owner (user)"
+    owner: User @goField(forceResolver: true)
 }
 
 """
@@ -175,6 +178,9 @@ input ContactInput {
 
     "A phone number associated with the contact."
     phoneNumber: PhoneNumberInput
+
+    "Id of the contact owner (user)"
+    ownerId: ID
 }
 
 """
@@ -212,6 +218,9 @@ input ContactUpdateInput {
 
     "User-defined field that defines the relationship type the contact has with your business.  `Customer`, `Partner`, `Lead` are examples."
     contactTypeId: ID
+
+    "Id of the contact owner (user)"
+    ownerId: ID
 }
 
 """
